Refactored.
All tests passing.

+++ src/Euroleague/Series1/ISeries1Case.sol
@@ -6,5 +6,5 @@ interface ISeries1Case {
     error MomentsContractNotValid();
     error MomentsPerCaseNotValid();
 
-    event CaseOpened(uint96 indexed token_id, address indexed case_address);
+    event CaseOpened(uint256 indexed token_id, address indexed case_address);
 }
